backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
PyJWT==2.10.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
google-generativeai==0.8.3
